fix nb args (overflow)

+++ src/ast.c
@@ -83,7 +83,7 @@ AST_NODE *AST_NODE_FUNC_DEC_Create(HashTable *namespace, AST_NODE *data) {
   return (AST_NODE *)ret;
 }
 
-AST_NODE *AST_NODE_FUNC_CALL_Create(AST_NODE *function, size_t nb_args) {
+AST_NODE *AST_NODE_FUNC_CALL_Create(AST_NODE *function, uint64_t nb_args) {
   AST_NODE_FUNC_CALL *ret = malloc(sizeof(struct AST_NODE_FUNC_CALL));
   assert(ret);
   ret->type = AST_NODE_TYPE_FUNC_CALL;
